Rename backup package to jetpack
