+++ WordPress/src/main/java/org/wordpress/android/ui/jetpack/backup/BackupProgressChecker.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.ui.backup.download
+package org.wordpress.android.ui.jetpack.backup
 
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.delay
