+++ WordPress/src/test/java/org/wordpress/android/ui/jetpack/RewindProgressCheckerTest.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.ui.activitylog
+package org.wordpress.android.ui.jetpack
 
 import com.nhaarman.mockitokotlin2.any
 import com.nhaarman.mockitokotlin2.inOrder
@@ -23,7 +23,7 @@ import org.wordpress.android.fluxc.store.ActivityLogStore
 import org.wordpress.android.fluxc.store.ActivityLogStore.OnRewindStatusFetched
 import org.wordpress.android.fluxc.store.ActivityLogStore.RewindStatusError
 import org.wordpress.android.fluxc.store.ActivityLogStore.RewindStatusErrorType.GENERIC_ERROR
-import org.wordpress.android.ui.backup.rewind.RewindProgressChecker
+import org.wordpress.android.ui.jetpack.rewind.RewindProgressChecker
 import java.util.Date
 
 @RunWith(MockitoJUnitRunner::class)
